add a ut case
